+++ src/components/user-interaction/MultiSelect/MultiSelectRoot.tsx
@@ -155,6 +155,7 @@ export function MultiSelectRoot<T>({
       highlightNext: state.highlightNext,
       highlightPrevious: state.highlightPrevious,
       highlightItem: state.highlightItem,
+      handleTypeaheadKey: state.handleTypeaheadKey,
       setIsOpen: state.setIsOpen,
       toggleIsOpen: state.toggleOpen,
       config: {